Update utils.py
Used FP16 for function inputs

+++ utils/utils.py
@@ -185,7 +185,7 @@ def bbox_wh_iou(wh1, wh2):
     wh2 = wh2.t()
     w1, h1 = wh1[0], wh1[1]
     w2, h2 = wh2[0], wh2[1]
-    inter_area = torch.min(w1, w2) * torch.min(h1, h2)
+    inter_area = torch.min(w1.float(), w2.float()) * torch.min(h1.float(), h2.float())
     union_area = (w1 * h1 + 1e-16) + w2 * h2 - inter_area
     return inter_area / union_area
